requirements.txt
boto3
streamlit
openpyxl
requests
mysql-connector-python
git+https://github.com/Oxford-Data-Processes/aws-utils.git
